pyproject.toml
[project]
name = "langraph-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.2",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
